# app/src/resources/static/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: MoneyHarbor — Accounts API
  version: 1.0.0
servers:
  - url: http://localhost:8080
tags:
  - name: Accounts
  - name: User

security:
  - BasicAuth: [ ]

paths:
  /accounts:
    post:
      tags: [ Accounts ]
      summary: Create account
      description: |
        Создаёт счёт. `title` необязателен; если `null`, сервер сгенерирует `"{currency}_счет"`.
        При невалидном JSON (ошибка десериализации) — 400.
      operationId: createAccount
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateAccountRequest' }
            examples:
              okWithoutTitle:
                summary: Без title (будет сгенерирован)
                value: { currency: "RUB", amount: 1500.50 }
              okWithTitle:
                summary: С указанным title
                value: { title: "Зарплатный", currency: "USD", amount: 0 }
      responses:
        '201':
          description: Created
          headers:
            Location:
              description: URL созданного ресурса
              schema: { type: string, format: uri }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountResponse' }
        '400':
          description: Bad Request (RFC 9457 Problem Details)
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetails' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

  /accounts/{id}:
    get:
      tags: [ Accounts ]
      summary: Get account by id
      operationId: getAccountById
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountResponse' }
        '400':
          description: Bad Request (invalid UUID) — RFC 9457 Problem Details
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetails' }
        '404':
          description: Not Found — RFC 9457 Problem Details
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetails' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

    delete:
      tags: [ Accounts ]
      summary: Delete account
      operationId: deleteAccount
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: No Content — запись существовала и была удалена
        '200':
          description: OK — записи не было (идемпотентно)
        '400':
          description: Bad Request (invalid UUID) — RFC 9457 Problem Details
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetails' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

  /accounts/{id}/title:
    patch:
      tags: [ Accounts ]
      summary: Update account title
      description: |
        Обновляет название счета. Доступно только пользователям с ролью USER.
        
        Валидация:
        - id должен быть валидным UUID
        - title обязателен (не null)
        - title.length <= 50
      operationId: updateAccountTitle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID счета
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAccountTitleRequest'
            examples:
              valid:
                summary: Валидный запрос
                value: { title: "Зарплатный счет" }
              nullTitle:
                summary: Title = null (невалидно)
                value: { title: null }
              longTitle:
                summary: Title > 50 символов (невалидно)
                value: { title: "Очень длинное название счета которое превышает лимит в 50 символов" }
      responses:
        '204':
          description: Title успешно обновлен
        '400':
#          $ref: '#/components/schemas/ProblemDetails'
          description: |
            Bad Request - RFC 9457 Problem Details
            Возможные причины:
            - Невалидный UUID формата
            - Title = null
            - Title.length > 50
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetails' }
#              examples:
#                invalidUuid:
#                  summary: Невалидный UUID
#                  value:
#                    type: "about:blank"
#                    title: "Bad Request"
#                    status: 400
#                    detail: "Invalid UUID format: 123"
#                    instance: "/accounts/123/title"
#                nullTitle:
#                  summary: Title не может быть null
#                  value:
#                    type: "about:blank"
#                    title: "Bad Request"
#                    status: 400
#                    detail: "title: must not be null"
#                    instance: "/accounts/3fa85f64-5717-4562-b3fc-2c963f66afa6/title"
#                longTitle:
#                  summary: Title превышает 50 символов
#                  value:
#                    type: "about:blank"
#                    title: "Bad Request"
#                    status: 400
#                    detail: "title: length must be between 1 and 50"
#                    instance: "/accounts/3fa85f64-5717-4562-b3fc-2c963f66afa6/title"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Not Found — RFC 9457 Problem Details
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetails' }
#              examples:
#                accountNotFound:
#                  summary: Счет не найден
#                  value:
#                    type: "about:blank"
#                    title: "Not Found"
#                    status: 404
#                    detail: "Account not found with id: 3fa85f64-5717-4562-b3fc-2c963f66afa6"
#                    instance: "/accounts/3fa85f64-5717-4562-b3fc-2c963f66afa6/title"

  /user/sign-up:
    post:
      tags: [ User ]
      summary: Регистрация пользователя
      description: Создаёт нового пользователя с ролью USER. Эндпоинт публичный.
      security: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserSignUpRequest' }
            examples:
              ok:
                value: { email: "alice@example.com", password: "Abcdefg1" }
              bad:
                value: { email: "not-an-email", password: "vvvvalera123" }
      responses:
        '201':
          description: Пользователь зарегистрирован
          headers:
            Location:
              description: URI созданного пользователя
              schema: { type: string, format: uri }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserResponse' }
              examples:
                created:
                  value:
                    id: "3f5b4e7a-2c1e-4b7a-9432-6a0b7a5b9e10"
                    email: "alice@example.com"
        '400':
          description: Ошибка валидации
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetails' }
              examples:
                validation:
                  value:
                    type: "about:blank"
                    title: "Invalid request"
                    status: 400
                    detail: "password: must match ^(?=.*\\d)(?!.*(.)\\1{2}).{8,}$; email: must be a valid email"
        '409':
          description: Пользователь с таким email уже существует
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetails' }
              examples:
                conflict:
                  value:
                    type: "about:blank"
                    title: "Conflict"
                    status: 409
                    detail: "Email already registered: alice@example.com"
  /user/profile:
    get:
      tags: [ User ]
      operationId: getUserProfile
      summary: Получение информации о профиле пользователя
      description: Получение информации о профиле пользователя, включает в себя информацию о пользователе и его счетах
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserProfileResponse' }

components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic

  responses:
    Unauthorized:
      description: Unauthorized — требуется аутентификация
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
          examples:
            unauthorized:
              value:
                type: "about:blank"
                title: "Unauthorized"
                status: 401
                detail: "Missing or invalid Authorization header"
    Forbidden:
      description: Forbidden — недостаточно прав
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
          examples:
            forbidden:
              value:
                type: "about:blank"
                title: "Forbidden"
                status: 403
                detail: "Access is denied"
  schemas:
    Currency:
      type: string
      description: Код валюты по ISO 4217.
      enum: [ BYN, RUB, USD ]
      x-enum-descriptions:
        - "Белорусский рубль"
        - "Российский рубль"
        - "Доллар США"
      example: RUB

    UserAccountDtoResponse:
      type: object
      required: [ currency, title, amount ]
      properties:
        title:
          type: string
          description: Если null — сервер сгенерирует "{currency}_счет".
          example: RUB_счет
        currency:
          $ref: '#/components/schemas/Currency'
        amount:
          type: number
          x-java-type: java.math.BigDecimal
          description: Денежная сумма; соответствует NUMERIC(19,2).
          example: 0
      example:
        title: RUB_счет
        currency: RUB
        amount: 0


    CreateAccountRequest:
      type: object
      required: [ currency ]
      properties:
        title:
          type: string
          nullable: true
          maxLength: 50
          description: Если null — сервер сгенерирует "{currency}_счет".
          example: null
        currency:
          $ref: '#/components/schemas/Currency'
        amount:
          type: number
          x-java-type: java.math.BigDecimal
          description: Денежная сумма; соответствует NUMERIC(19,2).
          example: 0
          multipleOf: 0.01
          maximum: 100000000000000000
          exclusiveMaximum: true
          minimum: -100000000000000000
          exclusiveMinimum: true
          x-field-extra-annotation: |
            @Digits(integer = 17, fraction = 2)
      example:
        title: null
        currency: RUB
        amount: 0

    AccountResponse:
      type: object
      required: [ title, currency, amount ]
      properties:
        title:
          type: string
        currency:
          $ref: '#/components/schemas/Currency'
        amount:
          type: number
          x-java-type: java.math.BigDecimal
          description: Денежная сумма; соответствует NUMERIC(19,2).
          example: 1500.50

    ProblemDetails:
      type: object
      description: RFC 9457 Problem Details object
      properties:
        type:
          type: string
          format: uri
          description: URI-идентификатор типа проблемы (или "about:blank")
          example: about:blank
        title:
          type: string
          description: Краткий заголовок проблемы
          example: Bad Request
        status:
          type: integer
          format: int32
          description: HTTP-статус для этой ошибки
          example: 400
        detail:
          type: string
          description: Человеко-читаемое пояснение
          example: Invalid value in request body
        instance:
          type: string
          format: uri
          description: URI конкретного случая (обычно путь запроса)
          example: /accounts/00000000-0000-0000-0000-000000000000
      additionalProperties: true
    UserSignUpRequest:
      type: object
      required: [ email, password ]
      properties:
        email:
          type: string
          format: email
          minLength: 1
          description: Логин в формате email
        password:
          type: string
          format: password
          minLength: 8
          description: >
            Пароль: длина ≥ 8, хотя бы одна цифра, без трёх одинаковых символов подряд
            символов (например, "vvv" запрещено).
          pattern: '^(?=.*\d)(?!.*(.)\1{2}).{8,}$'
          example: Abcdefg1
          x-pattern-message: 'Пароль: ≥ 8, хотя бы одна цифра, без трёх одинаковых символов подряд'
    UserProfileResponse:
      type: object
      required: [ email, accounts ]
      properties:
        email:
          type: string
          format: email
          description: Логин в формате email
        accounts:
          type: array
          items:
            $ref: "#/components/schemas/UserAccountDtoResponse"
          description: Информация о счетах пользователя

    UserResponse:
      type: object
      required: [ id, email ]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email

    UpdateAccountTitleRequest:
      type: object
      required: [ title ]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 50
          description: Новое название счета. Обязательное поле, длина от 1 до 50 символов.
          example: "Зарплатный счет"